refactor(react-doctor): decompõe CommentCard e LlmInsightsView (Stats, #339 Fase 3)
Extrai subcomponentes/hooks preservando comportamento até os dois
componentes saírem do threshold de no-giant-component:

- CommentCard (314 → ~105 linhas): CommentCardHeader, GabaritoSection
  (com estado/fetch do gabarito), SuggestionActions, ExclusionActions
  (arquivo próprio) e comment-card-utils.ts (labels/formatters).
- LlmInsightsView (396 → ~120 linhas): useLlmErrorFiltering (funde o
  estado dos filtros com toda a derivação — população filtrada, taxa
  de erro, ordenação, contagens), ErrorStatsCards e ErrorFiltersToolbar.

O override de no-event-handler em doctor.config.json acompanha o código
movido (CommentCard.tsx → ExclusionActions.tsx) — relocação de débito
já suprimido, não supressão nova.

react-doctor: 16 → 14 diagnósticos (restam apenas pré-existentes).
vitest 863/863; typecheck sem erros novos (18 pré-existentes em
fixtures de ComparePage, alheios ao PR).

Refs #355
